[Refactor:Globals] Consolidate global interface definitions and remove unused properties
